# cmd/server/main.go
package main

import (
	"context"
	"fmt"
	"net/http"
	"os"
	"os/signal"
	"strings"
	"syscall"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/redis/go-redis/v9"
	"github.com/spf13/viper"
	"go.uber.org/zap"
	"gorm.io/gorm"

	"github.com/trpg-solo-engine/backend/internal/domain"
	"github.com/trpg-solo-engine/backend/internal/handler"
	"github.com/trpg-solo-engine/backend/internal/infrastructure/database"

	// "github.com/trpg-solo-engine/backend/internal/middleware"
	"github.com/trpg-solo-engine/backend/internal/service"
)

func main() {
	// 初始化日志
	logger, err := zap.NewProduction()
	if err != nil {
		panic(fmt.Sprintf("failed to initialize logger: %v", err))
	}
	defer logger.Sync()

	// 加载配置
	if err := loadConfig(); err != nil {
		logger.Fatal("failed to load config", zap.Error(err))
	}

	// 初始化数据库
	db, err := database.NewPostgresDB(logger)
	if err != nil {
		logger.Fatal("failed to initialize database", zap.Error(err))
	}

	// 运行数据库迁移
	if err := database.RunMigrations(db, logger); err != nil {
		logger.Fatal("failed to run migrations", zap.Error(err))
	}

	// 初始化Redis
	redisClient, err := database.NewRedisClient(logger)
	if err != nil {
		logger.Fatal("failed to initialize redis", zap.Error(err))
	}
	defer redisClient.Close()

	// 初始化服务
	diceService := domain.NewDiceService()
	agentService := service.NewAgentService()
	gameService := service.NewGameService()
	scenarioService := service.NewScenarioService("scenarios")
	saveService := service.NewSaveService(gameService, agentService)

	// 创建Gin路由
	router := setupRouter(logger, db, redisClient, diceService, agentService, gameService, scenarioService, saveService)

	// 创建HTTP服务器
	port := viper.GetString("server.port")
	if port == "" {
		port = "8080"
	}

	srv := &http.Server{
		Addr:    ":" + port,
		Handler: router,
	}

	// 启动服务器
	go func() {
		logger.Info("starting server", zap.String("port", port))
		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			logger.Fatal("failed to start server", zap.Error(err))
		}
	}()

	// 优雅关闭
	quit := make(chan os.Signal, 1)
	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)
	<-quit

	logger.Info("shutting down server...")

	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()

	if err := srv.Shutdown(ctx); err != nil {
		logger.Fatal("server forced to shutdown", zap.Error(err))
	}

	logger.Info("server exited")
}

func loadConfig() error {
	viper.SetConfigName("config")
	viper.SetConfigType("yaml")
	viper.AddConfigPath(".")
	viper.AddConfigPath("./configs")

	// 设置默认值
	viper.SetDefault("server.port", "8080")
	viper.SetDefault("log.level", "info")
	viper.SetDefault("database.host", "localhost")
	viper.SetDefault("database.port", 5432)
	viper.SetDefault("database.user", "trpg")
	viper.SetDefault("database.password", "trpg_password")
	viper.SetDefault("database.dbname", "trpg_solo_engine")
	viper.SetDefault("database.sslmode", "disable")
	viper.SetDefault("redis.host", "localhost")
	viper.SetDefault("redis.port", 6379)

	// 启用环境变量支持
	viper.AutomaticEnv()
	// 设置环境变量前缀（可选）
	// viper.SetEnvPrefix("TRPG")
	// 替换环境变量中的点号为下划线
	viper.SetEnvKeyReplacer(strings.NewReplacer(".", "_"))

	if err := viper.ReadInConfig(); err != nil {
		if _, ok := err.(viper.ConfigFileNotFoundError); ok {
			// 配置文件不存在，使用默认值和环境变量
			return nil
		}
		return err
	}

	return nil
}

func setupRouter(logger *zap.Logger, db *gorm.DB, redisClient *redis.Client, diceService domain.DiceService, agentService service.AgentService, gameService service.GameService, scenarioService service.ScenarioService, saveService service.SaveService) *gin.Engine {
	// 设置Gin模式
	if viper.GetString("server.mode") == "release" {
		gin.SetMode(gin.ReleaseMode)
	}

	router := gin.New()

	// 添加中间件
	router.Use(gin.Recovery())
	// 使用简单的内联日志中间件（可以替换为 middleware.LoggerMiddleware(logger)）
	router.Use(loggerMiddleware(logger))

	// 可选：使用更强大的日志中间件
	// router.Use(middleware.LoggerMiddleware(logger))
	// router.Use(middleware.PerformanceTrackerMiddleware(logger))

	// 健康检查端点
	router.GET("/health", func(c *gin.Context) {
		// 检查数据库连接
		sqlDB, err := db.DB()
		dbStatus := "ok"
		if err != nil || sqlDB.Ping() != nil {
			dbStatus = "error"
		}

		// 检查Redis连接
		redisStatus := "ok"
		if err := redisClient.Ping(c.Request.Context()).Err(); err != nil {
			redisStatus = "error"
		}

		c.JSON(http.StatusOK, gin.H{
			"status":   "ok",
			"service":  "trpg-solo-engine",
			"database": dbStatus,
			"redis":    redisStatus,
		})
	})

	// 初始化处理器
	diceHandler := handler.NewDiceHandler(diceService, agentService)
	agentHandler := handler.NewAgentHandler(agentService)
	sessionHandler := handler.NewSessionHandler(gameService)
	scenarioHandler := handler.NewScenarioHandler(scenarioService)
	saveHandler := handler.NewSaveHandler(saveService, gameService)

	// API文档路由
	router.GET("/api/docs", func(c *gin.Context) {
		c.Redirect(http.StatusMovedPermanently, "/api/docs/")
	})
	router.GET("/api/docs/", func(c *gin.Context) {
		c.File("api/swagger-ui.html")
	})
	router.GET("/api/docs/openapi.yaml", func(c *gin.Context) {
		c.File("api/openapi.yaml")
	})

	// API路由组
	api := router.Group("/api")
	{
		api.GET("/version", func(c *gin.Context) {
			c.JSON(http.StatusOK, gin.H{
				"version": "0.1.0",
				"name":    "TRPG Solo Engine",
			})
		})

		// 骰子API
		dice := api.Group("/dice")
		{
			dice.POST("/roll", diceHandler.RollDice)
			dice.POST("/ability", diceHandler.RollForAbility)
			dice.POST("/request", diceHandler.RollForRequest)
		}

		// 角色API
		agents := api.Group("/agents")
		{
			agents.POST("", agentHandler.CreateAgent)
			agents.GET("", agentHandler.ListAgents)
			agents.GET("/:id", agentHandler.GetAgent)
			agents.PUT("/:id", agentHandler.UpdateAgent)
			agents.DELETE("/:id", agentHandler.DeleteAgent)
		}

		// 游戏会话API
		sessions := api.Group("/sessions")
		{
			sessions.POST("", sessionHandler.CreateSession)
			sessions.GET("/:id", sessionHandler.GetSession)
			sessions.POST("/:id/actions", sessionHandler.ExecuteAction)
			sessions.POST("/:id/phase", sessionHandler.TransitionPhase)
		}

		// 剧本API
		scenarios := api.Group("/scenarios")
		{
			scenarios.GET("", scenarioHandler.ListScenarios)
			scenarios.GET("/:id", scenarioHandler.GetScenario)
			scenarios.GET("/:id/scenes/:sceneId", scenarioHandler.GetScene)
		}

		// 存档API
		saves := api.Group("/saves")
		{
			saves.POST("", saveHandler.CreateSave)
			saves.GET("", saveHandler.ListSaves)
			saves.GET("/:id", saveHandler.GetSave)
			saves.POST("/:id/load", saveHandler.LoadSave)
			saves.DELETE("/:id", saveHandler.DeleteSave)
		}
	}

	return router
}

func loggerMiddleware(logger *zap.Logger) gin.HandlerFunc {
	return func(c *gin.Context) {
		start := time.Now()
		path := c.Request.URL.Path
		query := c.Request.URL.RawQuery

		c.Next()

		latency := time.Since(start)

		logger.Info("request",
			zap.String("method", c.Request.Method),
			zap.String("path", path),
			zap.String("query", query),
			zap.Int("status", c.Writer.Status()),
			zap.Duration("latency", latency),
			zap.String("ip", c.ClientIP()),
		)
	}
}